[Metal] Fix signed/unsigned compare

+++ aten/src/ATen/native/metal/mpscnn/MPSCNNConvOp.mm
@@ -75,10 +75,10 @@
   using namespace at::native::metal::mpscnn;
   TORCH_CHECK(
       params.DX == params.DY == 1, "Dilated convolution is not supported yet.");
-  const int64_t oC = params.OC;
-  const int64_t iC = params.C;
-  const int64_t kH = params.KH;
-  const int64_t kW = params.KW;
+  const NSUInteger oC = params.OC;
+  const NSUInteger iC = params.C;
+  const NSUInteger kH = params.KH;
+  const NSUInteger kW = params.KW;
   MPSCNNNeuron* neuron = at::native::metal::neuron(t);
   MPSCNNConvolutionDescriptor* desc = nil;
   if (params.isDepthwise()) {
